requirements.txt
streamlit
fpdf2
Pillow
pymupdf